Completed enemy Wave Spawner Script
At least I hope it's finished
I added some more stuff like
The queue system to prevent more than a specific amount of enemies to be alive at once in a wave.
I also added some checks to ensure smoothness and no weird errors
I added some debugging keys to make sure things work the way I want them to work (need to remove these in the final build)
I added a list of waves that would include a mini boss, a specific wave (that can be changed in the inspector) for the final boss and so on.